final commit

+++ player.java
@@ -1,5 +1,5 @@
 public class player {
-    int money;
+    double money;
     int items;
     int person;
 
@@ -7,11 +7,11 @@ public class player {
         if (person == 1){
             money = 25000000;
         }else if(person == 2){
-            money = 20000000;
+            money = 20000000.0;
         }else if (person == 3){
-            money = 600000000;
+            money = 600000000.0;
         }else{
-            money = 5000000;
+            money = 5000000.0;
         }
     }
 }
